fix: cast token0/1Exposure to uint from PanopticMath output, and cast them into int as a PanopticMath input

+++ src/accountants/PanopticVaultAccountant.sol
@@ -230,7 +230,7 @@ contract PanopticVaultAccountant is Ownable {
                 );
 
                 poolExposure0 = PanopticMath.convert0to1(poolExposure0, conversionPrice);
-                token0Exposure = PanopticMath.convert0to1(token0Exposure, conversionPrice);
+                token0Exposure = uint256(PanopticMath.convert0to1(int256(token0Exposure), conversionPrice));
             }
 
             if (address(pools[i].token1) != underlyingToken) {
@@ -248,7 +248,7 @@ contract PanopticVaultAccountant is Ownable {
                 );
 
                 poolExposure1 = PanopticMath.convert1to0(poolExposure1, conversionPrice);
-                token1Exposure = PanopticMath.convert1to0(token1Exposure, conversionPrice);
+                token1Exposure = uint256(PanopticMath.convert1to0(int256(token1Exposure), conversionPrice));
             }
 
             // debt in pools with negative exposure does not need to be paid back
